+++ demos/angular/tinymce5/src/app/app.component.ts
@@ -36,8 +36,6 @@ export class AppComponent {
 
   // Set options for the editor.
   public options: Object = {
-    height: 500,
-    menubar: false,
     base_url: '/tinymce', // Root for resources
     suffix: '.min',        // Suffix to use when loading resources
 
@@ -45,17 +43,21 @@ export class AppComponent {
     external_plugins: {
       'tiny_mce_wiris' : `${window.location.href}/node_modules/@wiris/mathtype-tinymce5/plugin.min.js`
     },
-    plugins: [
-      'advlist autolink lists link image charmap print preview anchor',
-      'searchreplace visualblocks code fullscreen',
-      'insertdatetime media table paste code help wordcount '
-    ],
-    toolbar: [
-      ' bold italic |' +
-       'tiny_mce_wiris_formulaEditor tiny_mce_wiris_formulaEditorChemistry '
-    ],
     htmlAllowedTags:  ['.*'],
     htmlAllowedAttrs: ['.*'],
+
+    // We recommend to set 'draggable_modal' to true to avoid overlapping issues
+    // with the different UI modal dialog windows implementations between core and third-party plugins on TinyMCE.
+    // @see: https://github.com/wiris/html-integrations/issues/134#issuecomment-905448642
+    draggable_modal: true,
+    plugins: ['image', 'media'],
+    toolbar: 'undo redo | styleselect | bold italic | image media | tiny_mce_wiris_formulaEditor tiny_mce_wiris_formulaEditorChemistry',
+
+    // language: 'fr_FR',
+    // You could set a different language for MathType editor:
+    // mathTypeParameters: {
+    //   editorParameters: { language: 'de' },
+    // },
 
     // Handle events.
     setup(editor) {