+++ demos/html5/tinymce4/src/app.js
@@ -17,7 +17,8 @@ tinymce.init({                                                                  
   external_plugins: {
     tiny_mce_wiris: `${window.location.href}node_modules/@wiris/mathtype-tinymce4/plugin.min.js`,
   },
-  toolbar: 'undo redo | styleselect | bold italic | tiny_mce_wiris_formulaEditor tiny_mce_wiris_formulaEditorChemistry',
+  plugins: ['image'],
+  toolbar: 'undo redo | styleselect | bold italic| image | tiny_mce_wiris_formulaEditor tiny_mce_wiris_formulaEditorChemistry',
 
   // language: 'fr_FR',
   // // You could set a different language for MathType editor: